Updated root layout architicture for dynamically scaling
