+++ src/ElectronApp.ts
@@ -113,7 +113,7 @@ export default class ElectronApp {
     this.window = new electron.BrowserWindow({
       width: 800,
       height: 700,
-      maxWidth: 1000,
+      maxWidth: 1400,
       minWidth: 525,
       minHeight: 325,
       icon: path.join(__dirname, 'images/icon-64x64.png'),
